embedded all boost::thread stuff inside Thread
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@7792 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/c++/util/Timeout.cpp
@@ -232,7 +232,7 @@ namespace org { namespace xmlBlaster { namespace util {
          {
             log_.trace(ME, "sleeping ... " + lexical_cast<string>(delay) + " milliseconds");
             isReady_ = true;
-            waitForTimeoutCondition_.wait(waitForTimeoutLock, delay);
+            waitForTimeoutCondition_.wait(waitForTimeoutLock, (long)delay);
             log_.trace(ME, "waking up .. ");
          }
       }
